rename several files for the dart backend
