+++ generator/internal/dart/annotate_test.go
